Move conversion of Data16-to-Data32 into set_N_bits

+++ CmdProcessor/include/adf4356.h
@@ -87,7 +87,8 @@ class ADF4356_LO {
     
     // ADF4356 methods
     void begin(uint8_t);
-    void set_N_bits(uint32_t);
+    void set_N_bits(uint16_t);
+    // void set_N_bits(uint32_t);
     uint32_t turn_off_RF();
     uint32_t set_n4dBm();
     uint32_t set_n1dBm();
